Cnahe the BigInteger to BigDecimal - typo in last patch
git-svn-id: http://svn.codehaus.org/groovy/trunk@1351 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ groovy/modules/xmlrpc/src/main/uk/co/wilson/net/xmlrpc/XMLRPCMessageProcessor.java
@@ -40,7 +40,6 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.UnsupportedEncodingException;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -217,8 +216,8 @@ public class XMLRPCMessageProcessor extends MinML {
 				buffer.append("</struct></value>");
 			} else if (param instanceof GString) {	 // GString can be subclassed so the initial elements.get() can fail
 				((Emitter)elements.get(GString.class)).emit(buffer, param);
-			} else if (param instanceof BigInteger) {	 // BigInteger can be subclassed so the initial elements.get() can fail
-				((Emitter)elements.get(BigInteger.class)).emit(buffer, param);
+			} else if (param instanceof BigDecimal) {	 // BigDecimal can be subclassed so the initial elements.get() can fail
+				((Emitter)elements.get(BigDecimal.class)).emit(buffer, param);
 			} else if (param instanceof Date) {	 // Date can be subclassed so the initial elements.get() can fail
 				((Emitter)elements.get(Date.class)).emit(buffer, param);
 			} else {
